% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2sf.R
\name{coords2sf}
\alias{coords2sf}
\title{Convert coordinates to sfc_POINTS}
\usage{
coords2sf(lon, lat, crs)
}
\arguments{
\item{lon}{Numeric matrix with the longitude of each point}

\item{lat}{Numeric matrix with the latitude of each point}

\item{crs}{Integer or String. EPSG code or PROJ4 string of the points}
}
\value{
points as sfc_POINTS
}
\description{
This function converts normal coordinates in a grid configuration to
to sfc_POINTS type. Keep in mind that the first grid should be the
lower-left and the last one the upper-right (increasing longitude and
latitude).
}
