% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraCrop.R
\name{extraCrop}
\alias{extraCrop}
\title{Get extra bbox area for sf::st_crop}
\usage{
extraCrop(target, reference)
}
\arguments{
\item{target}{The target grid centroids.}

\item{reference}{Reference polygon used for the cropping.}
}
\value{
bbox
}
\description{
This function expands the st_crop boundary box by an specific amount. It is
useful if you want to construct a grid from just points. If you happen to
use only st_crop, you cannot get the full area that the grid would cover
because the crop would only reach the center of the grid. By adding thi
"specific amount" (half of the spatial resolution), all the grid area will
be covered.
}
