% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeights.R
\name{getWeights}
\alias{getWeights}
\title{Create grid mask for a polygon}
\usage{
getWeights(grid, shp)
}
\arguments{
\item{grid}{Grid (as polygons) to be intersected.}

\item{shp}{Polygons that will be used to intersect the grids.}
}
\value{
Weight for each cell index. NA means no intersection.
}
\description{
This function calculates the weight for each cell with respect to the area
that intersects with a polygon.
}
