% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridMaker.R
\name{gridMaker}
\alias{gridMaker}
\title{Create grid with polygons}
\usage{
gridMaker(grid, crs = 4326, export = FALSE, name = "malla.shp")
}
\arguments{
\item{grid}{Raster image or points (in a grid configuration, from lower-left
to uper-right).}

\item{crs}{Integer or String. EPSG code or PROJ4 string for polygons.}

\item{export}{Logic. Save the polygons in file.}

\item{name}{String. Name of the file to be exported.}
}
\value{
Grid as polygons
}
\description{
This function takes each grid from longitude and latitude vectors or a raster
image and creates polygons wich later can be used with sf package functions.
}
