% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskMaker.R
\name{maskMaker}
\alias{maskMaker}
\title{Create grid mask for a polygon}
\usage{
maskMaker(
  grid,
  shp,
  format = "array",
  key = NULL,
  atrib_c = NULL,
  export = NULL,
  weights = NULL,
  progress = TRUE
)
}
\arguments{
\item{grid}{Grid to be intersected.}

\item{shp}{Polygons that will be used to intersect the grids.}

\item{format}{Strimg. Output format.}

\item{key}{Numeric. Field that will be used to define the dimensions of output
files.}

\item{atrib_c}{String. Additional field to final table. Only for "list"
format.}

\item{export}{String. Name of the output file.}

\item{weights}{String. Name of the file with the weights.}

\item{progress}{Logic. Turn on progress bar.}
}
\value{
Array for every polygon intersected.
}
\description{
This function find all the grids that intersect a polygon and calculates the
fraction of the grids that are inside of the polygon.
}
